Fix matcens with multiple robot types never giving Algo weapons

+++ d2/main/gameseq.c
@@ -2726,10 +2726,10 @@ void examine_path_partime(point_seg* path, int path_count)
 											totalRobotTime += calculate_combat_time(NULL, robInfo, 0, 1);
 											if (robInfo->contains_type == OBJ_ROBOT) {
 												totalRobotTime += calculate_combat_time(NULL, &Robot_info[robInfo->contains_id], 0, 1) * round((robInfo->contains_count * (robInfo->contains_prob / 16)));
-												robotHasPowerup(robInfo->contains_id, (double)(1 / num_types));
+												robotHasPowerup(robInfo->contains_id, 1.0 / num_types);
 											}
 											else
-												robotHasPowerup(legal_types[n], (double)(1 / num_types));
+												robotHasPowerup(legal_types[n], 1.0 / num_types);
 										}
 										totalAmmoUsage += ParTime.ammo_usage;
 									}
